internal/transformer: Reduce some multi-value calc expressions
The calc reducer can now reduce expressions of the form
calc(100% - 1px - 1px), where the second and third values are the same
unit but the first isn't. It can also reduce calc(100% + 1px - 1px) to
calc(100%).

The reducer can no longer handle expressions of the form
calc(100% - 1px - -1px), where a negative value is subtracted. It also
can't reduce expressions where the first and third values are the same
type, but the second isn't.

Co-authored-by: Stephen Wan <stephen@stephenwan.net>